Projekte/Python/BioreactorRasPi-4a67154bc02fd2cfb791236cb316a456c9202c49/requirements.txt
matplotlib~=3.3.2
pyserial~=3.4
setuptools~=50.3.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib